Rescan CA certs when redirecting from HTTP to HTTPS
Closes #21636

Since the client CA certs are scanned lazily, the client must check and potentially load the CA certs any time it might connect to an HTTPS server, including while handling a redirect. To this end, this commit moves the logic for rescanning the CA certs right before the TLS client is initialized when establishing a connection.